requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login
Flask-Bcrypt
Flask-Migrate==4.0.5
python-docx==1.1.2
pdfplumber==0.7.6
pypdf==3.17.4
pytesseract==0.3.10
openai==0.28.1
Pillow==10.4.0
python-multipart==0.0.6
Werkzeug==2.2.3
requests==2.28.2
python-dotenv==0.21.1
psycopg2-binary==2.9.10

# Additional dependencies for advanced features
pandoc==2.3
markdown==3.5.1
weasyprint==60.2
reportlab==4.0.7
fpdf2==2.7.6
docx2pdf==0.1.8
markdown2==2.4.10
html2text==2020.1.16
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
jinja2==3.1.2
PyMuPDF==1.23.8

